Rename models path to allow each model to have their own file
<4582791+smcunning@users.noreply.github.com>
<65981543+carson-jardine@users.noreply.github.com>
<60277914+c-bartell@users.noreply.github.com>
