+++ api/database/models/encounters.py
