+++ api/resources/quests.py
@@ -8,11 +8,11 @@ from flask_restful import Resource, abort, reqparse, fields, marshal_with
 from sqlalchemy.orm.exc import NoResultFound
 
 from api import db
-from api.database.models_object.quests import Quest
-from api.database.models_object.users import User
-from api.database.models_object.user_quests import UserQuest
-from api.database.models_object.encounters import Encounter
-from api.database.models_object.actions import Action
+from api.database.models.quests import Quest
+from api.database.models.users import User
+from api.database.models.user_quests import UserQuest
+from api.database.models.encounters import Encounter
+from api.database.models.actions import Action
 
 def _quest_encounters_payload(encounter):
     return {
