+++ tests/endpoints/quests/test_get_encounters_by_quest.py
@@ -3,11 +3,11 @@ import unittest
 from unittest.mock import patch
 from copy import deepcopy
 from api import create_app, db
-from api.database.models_object.users import User
-from api.database.models_object.quests import Quest
-from api.database.models_object.user_quests import UserQuest
-from api.database.models_object.encounters import Encounter
-from api.database.models_object.actions import Action
+from api.database.models.users import User
+from api.database.models.quests import Quest
+from api.database.models.user_quests import UserQuest
+from api.database.models.encounters import Encounter
+from api.database.models.actions import Action
 from tests import db_drop_everything, assert_payload_field_type_value, \
     assert_payload_field_type
