+++ tests/endpoints/user_quests/test_get_quests_for_user.py
@@ -3,10 +3,9 @@ import unittest
 from unittest.mock import patch
 from copy import deepcopy
 from api import create_app, db
-from api.database.models_object.users import User
-from api.database.models_object.quests import Quest
-from api.database.models_object.user_quests import UserQuest
-# from api.database.models import UserQuest
+from api.database.models.users import User
+from api.database.models.quests import Quest
+from api.database.models.user_quests import UserQuest
 from tests import db_drop_everything, assert_payload_field_type_value, \
     assert_payload_field_type
