+++ tests/endpoints/users/test_get_user_by_email.py
@@ -4,8 +4,7 @@ from unittest.mock import patch
 from copy import deepcopy
 
 from api import create_app, db
-from api.database.models_object.users import User
-# from api.database.models import User
+from api.database.models.users import User
 from tests import db_drop_everything, assert_payload_field_type_value, \
     assert_payload_field_type
